+++ src/jvmMain/kotlin/com/github/dzivko1/dullcoin/domain/blockchain/usecase/GetBalanceUseCase.kt
@@ -6,5 +6,5 @@ import kotlinx.coroutines.flow.Flow
 class GetBalanceUseCase(
     private val blockchainService: BlockchainService
 ) {
-    operator fun invoke(): Flow<Int> = blockchainService.balanceFlow
+    operator fun invoke(): Flow<Long> = blockchainService.balanceFlow
 }
